// packages/common-next/types/account.ts
export type CleengCaptureField = {
  key: string;
  enabled: boolean;
  required: boolean;
  answer: string | Record<string, string | null> | null;
};

export type CleengCaptureQuestionField = {
  key: string;
  enabled: boolean;
  required: boolean;
  value: string;
  question: string;
  answer: string | null;
};
// ...
export type Customer = {
  id: string;
  email: string;
  firstName?: string;
  country?: string;
  metadata: Record<string, unknown>;
  lastName?: string;
  fullName?: string;
  [key: string]: unknown;
};
// ...
export type CustomerConsent = {
  customerId?: string;
  date?: number;
  label?: string;
  name: string;
  needsUpdate?: boolean;
  newestVersion?: string;
  required?: boolean;
  state: 'accepted' | 'declined';
  value?: string | boolean;
  version: string;
};

export type CustomerConsentArgs = {
  customer: Customer;
};

export type UpdateCustomerConsentsArgs = {
  customer: Customer;
  consents: CustomerConsent[];
};

export type GetCaptureStatusResponse = {
  isCaptureEnabled: boolean;
  shouldCaptureBeDisplayed: boolean;
  settings: Array<CleengCaptureField | CleengCaptureQuestionField>;
};

export type CaptureCustomAnswer = {
  questionId: string;
  question: string;
  value: string;
};

export type Capture = {
  firstName?: string;
  address?: string;
  address2?: string;
  city?: string;
  state?: string;
  postCode?: string;
  country?: string;
  birthDate?: string;
  companyName?: string;
  phoneNumber?: string;
  customAnswers?: CaptureCustomAnswer[];
};

export type GetCaptureStatusArgs = {
  customer: Customer;
};

export type UpdateCaptureStatusArgs = {
  customer: Customer;
} & Capture;

export type UpdateCaptureAnswersPayload = {
  customerId: string;
} & Capture;

export type FirstLastNameInput = {
  firstName: string;
  lastName: string;
  metadata?: Record<string, string | boolean>;
};

export type EmailConfirmPasswordInput = {
  email: string;
  confirmationPassword: string;
};

export type CommonAccountResponse = {
  message: string;
  code: number;
  errors?: Record<string, string>;
};

export type DeleteAccountPayload = {
  password: string;
};

export type SubscribeToNotificationsPayload = {
  uuid: string;
  onMessage: (payload: string) => void;
};